refactor(csrc): split profile collector into runtime stall summary units
